+++ src/Opaleye/Select.hs
@@ -1,4 +1,4 @@
--- | A 'Select' represents an SQL @SELECT@ statment.  To run a
+-- | A 'Select' represents an SQL @SELECT@ statement.  To run a
 -- 'Select' use the functions in "Opaleye.RunSelect".  To create a
 -- 'Select' you probably want to start by querying one of your
 -- 'Opaleye.Table.Table's using 'Opaleye.Table.selectTable'.