+++ src/Opaleye/Order.hs
@@ -75,14 +75,14 @@ desc = O.order HPQ.OrderOp { HPQ.orderDirection = HPQ.OpDesc
 
 -- | Specify an ascending ordering by the given expression.
 --   (Any NULLs appear first)
-ascNullsFirst :: SqlOrd b => (a -> F.Field b) -> O.Order a
+ascNullsFirst :: SqlOrd b => (a -> F.Field_ n b) -> O.Order a
 ascNullsFirst = O.order HPQ.OrderOp { HPQ.orderDirection = HPQ.OpAsc
                                     , HPQ.orderNulls     = HPQ.NullsFirst }
 
 
 -- | Specify an descending ordering by the given expression.
 --   (Any NULLs appear last)
-descNullsLast :: SqlOrd b => (a -> F.Field b) -> O.Order a
+descNullsLast :: SqlOrd b => (a -> F.Field_ n b) -> O.Order a
 descNullsLast = O.order HPQ.OrderOp { HPQ.orderDirection = HPQ.OpDesc
                                     , HPQ.orderNulls     = HPQ.NullsLast }